+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Platform Version 0.9.5 - UNRELEASED 
 * Update `Debug` printout for `SmartStreamWasm` to reduce noise ([#1524](https://github.com/infinyon/fluvio/pull/1524))
 * Increase platform stability ([#1497](https://github.com/infinyon/fluvio/pull/1497))
+* Spawn a thread to handle stream fetch requests ([#1522](https://github.com/infinyon/fluvio/issues/1522))
 
 ## Platform Version 0.9.4 - 2021-08-26 
 * Publish docker image for aarch64 #1389 ([#1389](https://github.com/infinyon/fluvio/pull/1389))